# Optimize model to leverage inplace to avoid create new tensor