Optimize the model a bit to reduce new tensor to created. This helps both memory and a bit runtime performanceas less cache conflict.

Before (latest main)
PS C:\gpt\llama.cpp> .\build\bin\RelWithDebInfo\main.exe -m C:\GPT\en-models\7B\ggml-alpaca-7b-q4.bin -t 16 --color -b 256 --top_k 10000 --temp 0.2 --repeat_penalty 1 -p 'Below is an instruction that describes a task. Write a response that appropriately completes the request. Tell me about alpacas.'
main: seed = 1680273128
llama_model_load: loading model from 'C:\GPT\en-models\7B\ggml-alpaca-7b-q4.bin' - please wait ...
llama_model_load: n_ctx   = 512
llama_model_load: n_mult  = 256
llama_model_load: n_head  = 32
llama_model_load: n_layer = 32
llama_model_load: n_rot   = 128
llama_model_load: f16     = 2
llama_model_load: n_ff    = 11008
llama_model_load: n_parts = 1
llama_model_load: type    = 1
llama_model_load: ggml map size = 4017.70 MB
llama_model_load: ggml ctx size =  81.25 KB
llama_model_load: mem required  = 5809.78 MB (+ 1026.00 MB per state)
llama_model_load: loading tensors from 'C:\GPT\en-models\7B\ggml-alpaca-7b-q4.bin'
llama_model_load: model size =  4017.27 MB / num tensors = 291
llama_init_from_file: kv self size  =  256.00 MB
system_info: n_threads = 16 / 20 | AVX = 1 | AVX2 = 0 | AVX512 = 0 | FMA = 0 | NEON = 0 | ARM_FMA = 0 | F16C = 0 | FP16_VA = 0 | WASM_SIMD = 0 | BLAS = 0 | SSE3 = 0 | VSX = 0 |
sampling: temp = 0.200000, top_k = 10000, top_p = 0.950000, repeat_last_n = 64, repeat_penalty = 1.000000
generate: n_ctx = 512, n_batch = 256, n_predict = 128, n_keep = 0


 Below is an instruction that describes a task. Write a response that appropriately completes the request. Tell me about alpacas. Alpacas are small, cud-chewing mammals that are related to camels. They are native to South America, where they have lived for thousands of years. Alpacas are highly social animals and live in herds of up to 20 individuals. They have a very good sense of hearing and can be quite vocal. Alpacas are shorn once a year, and their fleeces can be used to make a variety of products, such as clothing and home furnishings. They are also sometimes bred as pets. [end of text]

llama_print_timings:        load time = 11912.71 ms
llama_print_timings:      sample time =   416.96 ms /   121 runs   (    3.45 ms per run)
llama_print_timings: prompt eval time = 10468.02 ms /    29 tokens (  360.97 ms per token)
llama_print_timings:        eval time = 52864.24 ms /   120 runs   (  440.54 ms per run)
llama_print_timings:       total time = 65205.58 ms

After this change:
PS C:\gpt\llama.cpp> .\build\bin\RelWithDebInfo\main.exe -m C:\GPT\en-models\7B\ggml-alpaca-7b-q4.bin -t 16 --color -b 256 --top_k 10000 --temp 0.2 --repeat_penalty 1 -p 'Below is an instruction that describes a task. Write a response that appropriately completes the request. Tell me about alpacas.'
main: seed = 1680273585
llama_model_load: loading model from 'C:\GPT\en-models\7B\ggml-alpaca-7b-q4.bin' - please wait ...
llama_model_load: n_vocab = 32000
llama_model_load: n_ctx   = 512
llama_model_load: n_embd  = 4096
llama_model_load: n_mult  = 256
llama_model_load: n_head  = 32
llama_model_load: n_layer = 32
llama_model_load: n_rot   = 128
llama_model_load: f16     = 2
llama_model_load: n_ff    = 11008
llama_model_load: n_parts = 1
llama_model_load: type    = 1
llama_model_load: ggml map size = 4017.70 MB
llama_model_load: ggml ctx size =  81.25 KB
llama_model_load: mem required  = 5809.78 MB (+ 1026.00 MB per state)
llama_model_load: loading tensors from 'C:\GPT\en-models\7B\ggml-alpaca-7b-q4.bin'
llama_model_load: model size =  4017.27 MB / num tensors = 291
llama_init_from_file: kv self size  =  256.00 MB

system_info: n_threads = 16 / 20 | AVX = 1 | AVX2 = 0 | AVX512 = 0 | FMA = 0 | NEON = 0 | ARM_FMA = 0 | F16C = 0 | FP16_VA = 0 | WASM_SIMD = 0 | BLAS = 0 | SSE3 = 0 | VSX = 0 |
sampling: temp = 0.200000, top_k = 10000, top_p = 0.950000, repeat_last_n = 64, repeat_penalty = 1.000000
generate: n_ctx = 512, n_batch = 256, n_predict = 128, n_keep = 0


 Below is an instruction that describes a task. Write a response that appropriately completes the request. Tell me about alpacas. Alpacas are camelids native to South America. They are typically kept as pets, for their soft, luxurious fleece, which can be used to make coats, sweaters, and other textiles. Alpacas are social animals and live in herds with a herd leader. They are very curious and can be trained to follow basic commands. They are also quite gentle, and make excellent pets for families with children. Alpacas require a lot of space, however, so they are not ideal for small yards or apartments. [end of text]

llama_print_timings:        load time =  7973.50 ms
llama_print_timings:      sample time =   408.42 ms /   121 runs   (    3.38 ms per run)
llama_print_timings: prompt eval time =  7376.26 ms /    29 tokens (  254.35 ms per token)
llama_print_timings:        eval time = 50581.73 ms /   120 runs   (  421.51 ms per run)
llama_print_timings:       total time = 58975.67 ms