@slaren 2023-03-31
How does this work? From what I can tell, the `_inplace` versions of these ggml functions do exactly the same as their regular versions. Are the changes to `ggml.c` missing?