@slaren 2023-03-31
@howard0su in that case to avoid confusion I would suggest using only the `_inplace` versions of the functions that actually have inplace implementations. I suspect that the reason that they weren't exposed in `ggml.h` is because of that.