@howard0su 2023-03-31
> @howard0su in that case to avoid confusion I would suggest using only the `_inplace` versions of the functions that actually have inplace implementations. I suspect that the reason that they weren't exposed in `ggml.h` is because of that.

My preference is that if one operator has two versions, we should use inplace when possible. And we should fix ggml to have inplace and non-inplace version. because inplace and non-inplace is not determined by operator only. In case one tensor need to be used for two additional operators. Two operators have to be non-inplace. It is not big problem as only one model LLAMA is supported for now. but in a long run, we should be explict about that in both llama.cpp and ggml.c.

Torch API is a good reference if one API should support inplace or not.