@slaren 2023-03-31
I agree, it would be easy to introduce bugs if we are not very careful about what can be done inplace and what cannot. Ideally, we would have inplace implementations of all the functions that may benefit from it, but for now, if you think that we can gain some performance by using the currently available inplace implementations, let's use only these. Otherwise it is impossible to test if the conversion of the llama implementation to inplace functions has introduced any bugs.