@jart 2023-03-31
Thanks for taking the time to send this change!

As much as I'd love to support your work, my vote would be no. My testing indicates:

- Preserves same behavior (good)
- 0.8% decrease in overall clock cycles
- 4% performance improvement on evaluation (not good; beneath my 5% confidence threshold)
- 3% fewer page faults (not good; beneath my 5% confidence threshold)
- No change in peak resident memory usage (not good)
- Someone on our Discord reported that this PR causes startup to segfault for them

This change achieves its performance gains by introducing numerous new APIs that leverage an inherently risky, error prone, programming technique. I would feel comfortable doing those things if the advantages it gained us a 10x improvement, but for four percent it's probably not worth it, since that's below the confidence threshold of my ability to measure impact of changes.