@howard0su 2023-04-01
> No change in peak resident memory usage (not good)
This is due to the current implementation which allocate a fixed amount of memory which is not needed that much. I will evaluate it and make the change to the constant values in MEM_REQ_EVAL array.

> Someone on our Discord reported that this PR causes startup to segfault for them
Like to understand better on this. Let me join Discord to check.

> This change achieves its performance gains by introducing numerous new APIs that leverage an inherently risky, error prone, programming technique.
I don't agree with you. The original code is optimal to leverage inlace. If you look at transformer's implementation in torch, it is doing the exact thing I am proposing here to