@howard0su 2023-04-01
Here is some stats on memory usage drop:
(alpaca 7B model)
before:
llama_eval_internal: used_mem = 0.284 MB, scratch -- 4.393 MB 6.805 MB
after:
llama_eval_internal: used_mem = 0.272 MB, scratch -- 4.393 MB 3.809 MB

(llama 13B model)
before:
llama_eval_internal: used_mem = 0.355 MB, scratch -- 5.435 MB 8.531 MB
after:
llama_eval_internal: used_mem = 0.340 MB, scratch -- 4.927 MB 4.773 MB

this is not significant from the number. We only use several small tensors during eval and reuse the same buffer for every layer. but this reduce the speed of increasing when model is getting larger.