@howard0su 2023-04-01
About performance, I can conclude there is no obvious perf gain on my machine.

I use time of "per token" as reference here. with 7B alpaca model.
..\ministat.exe a.txt b.txt
x a.txt
+ b.txt
+--------------------------------------------------------------------------+
|           x      x        + +         +                                 x|
||_________________M______|___M_A__A___|_____________________________|     |
+--------------------------------------------------------------------------+
    N           Min           Max        Median           Avg        Stddev
x   3        239.21        252.07        240.62     243.96667     7.0530159
+   3        242.45           245        242.85     243.43333     1.3714348
No difference proven at 95.0% confidence

I strongly suggest to use this ministat tool to calculate the difference.