@ggerganov 2023-04-02
So my plan was to not expose the `_inplace` calls of `ggml` for the following reason:

When you use `_inplace` to create a computation graph, you immediately break the backward pass.
The purpose of the `_inplace` calls is to be used only by `ggml` internally during the backward pass computation to optimize some stuff where possible. Currently, the backward pass for LLaMA is not supported and is not being used.

Now, some of the `ggml` calls indeed implicitly perform `inplace` operations (e.g. `ggml_soft_max()`, `ggml_diag_mask_info()`, etc). The reason is mostly me being lazy to implement the non-inplace mode.

What would happen if we expose the `_inplace` API and use it in `llama.cpp`:

- We will now become explicit what is being inplace and what is not, which is good
- If in the future, we implement the backward pass for all operators, the `_inplace` calls in `llama.cpp` will have to be reverted, otherwise we won't be able to compute the backward pass. So this is not very good
- The API becomes more complicated. Also not very good

Technically, you could gain some efficiency when using inplace operators, given that you are interested only in doing forward pass inference. And in view of not implementing the backward pass for the remaining operators anytime soon, I guess it's not such a bad idea to use `_inplace` calls.

Still, I'm a little hesitant to make the change.
If you can demonstrate good performance benefit, we can merge it.
But at the moment, I feel like it's better the way it is.

In the meantime, will think about this some more