@howard0su 2023-04-02
Thank you for your explanation. It makes sense to me.

From the testing, the margin of performance gain is variable in different machine. In my 20cores old E5, the perf diff of per token time is very small. @jart  reported less than 5% gain in terms of cpu clock cycles and perf.

Not sure about your bar for good performance benefit. From what I tested so far, we may not easily get large perf gain at more. most low hang fruit is picked up.